backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
fastapi-cors
google-generativeai